refactored out formatter and html generator files
