+++ lib/formatter.js
@@ -1,4 +1,4 @@
-module.exports = {formatData}
+module.exports = {formatData, combineData}
 
 function formatData(dataCollection){
   var bittrexData = formatBittrex(dataCollection[0])
@@ -24,4 +24,26 @@ function formatPoloniex(data){
     return {name: "poloniex",
             bids: bids,
             asks: asks}
+}
+
+function combineData(data, orderType){
+  let prices = []
+  data.forEach(function(exchange){
+    exchange[orderType].forEach(function(order){
+      prices.push(order.rate)
+    })
+  })
+  let rowData = {}
+  let sortedPrices = null
+  if(orderType === "bids"){sortedPrices = prices.sort().reverse()}
+  if(orderType === "asks"){sortedPrices = prices.sort()}
+  sortedPrices.forEach(function(price){
+    rowData[price] = {"bittrex": 0, "poloniex": 0}
+  })
+  data.forEach(function(exchange){
+    exchange[orderType].forEach(function(order){
+      rowData[order.rate][exchange.name] += order.quantity
+    })
+  })
+  return rowData
 }
